+++ Makefile
@@ -6,7 +6,7 @@
 #    By: snicolet <snicolet@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2015/08/17 10:20:32 by snicolet          #+#    #+#              #
-#    Updated: 2017/10/23 16:41:09 by snicolet         ###   ########.fr        #
+#*   Updated: 2017/10/26 03:11:27 by snicolet         ###   ########.fr       *#
 #                                                                              #
 # **************************************************************************** #
 
@@ -83,7 +83,9 @@ MEMORY=ft_memset.o \
 	   ft_memdel.o \
 	   ft_memdup.o \
 	   ft_realloc.o \
-	   ft_malloc.o
+	   ft_malloc.o \
+	   ft_page.o \
+	   ft_block.o
 
 PRINTF_DIR=printf
 PRINTF=ft_printf.o \
